34: separate files and add completions object
