+++ rag-toolchain/src/clients/mod.rs
@@ -2,3 +2,4 @@
 pub mod openai_client;
 /// Module that defines the traits used by the clients in this crate
 pub mod traits;
+pub mod model;
