+++ rag-toolchain/src/clients/model/mod.rs
@@ -1,0 +1,2 @@
+pub mod chat_completions;
+pub mod embeddings;
